offline: extract everything to the offline package
To make room for the upcoming online mode, create the offline package.
Move all code there. Make the main file minimal, with all commands
definitions also moved to the offline package. Move integration tests to
the package, since they will have nothing to do with the upcoming online
mode.

Add `offline mode` category label to commands, so they're groupped
together when the help is printed.

This commit just shuffles files around, without adding or removing any
functionality.


ofline: missing command categories
