+++ .gitignore
@@ -1,7 +1,6 @@
 .idea/
 
 *.ph1
-*.ptau
 *.ph2
 
 trusted-setup
