--- cmd/doc.go
@@ -1,3 +1,0 @@
-// Package cmd provides handlers for the command line interface for the application.
-// The handlers are used in `main.go`.
-package cmd
