+++ offline/actions/contribute.go
@@ -1,4 +1,4 @@
-package cmd
+package actions
 
 import (
 	"context"
@@ -9,7 +9,7 @@ import (
 
 	"github.com/urfave/cli/v3"
 
-	"github.com/reilabs/trusted-setup/phase2"
+	"github.com/reilabs/trusted-setup/offline/phase2"
 )
 
 // appendTimestamp appends a timestamp to the given string if not present, or updates the timestamp if present.
