+++ offline/actions/doc.go
@@ -1,0 +1,3 @@
+// Package actions provides handlers for the command line interface for the application.
+// The handlers are used in `main.go`.
+package actions
