--- test/integration_test.go
@@ -1,9 +1,0 @@
-package test
-
-import (
-	"testing"
-)
-
-func TestIntegration(t *testing.T) {
-	t.Run("Test offline ceremony", TestOfflineCeremony)
-}
